+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/jni/ExpectedType.kt
@@ -33,6 +33,66 @@ class SingleType(
    */
   @DoNotStrip
   fun getSecondParameterType() = parameterTypes?.get(1)
+
+  override fun equals(other: Any?): Boolean {
+    if (this === other) {
+      return true
+    }
+
+    if (javaClass != other?.javaClass) {
+      return false
+    }
+
+    other as SingleType
+
+    if (expectedCppType != other.expectedCppType) {
+      return false
+    }
+    if (!parameterTypes.contentEquals(other.parameterTypes)) {
+      return false
+    }
+
+    return true
+  }
+
+  override fun hashCode(): Int {
+    var result = expectedCppType.hashCode()
+    result = 31 * result + (parameterTypes?.contentHashCode() ?: 0)
+    return result
+  }
+
+  companion object {
+    fun merge(
+      first: SingleType,
+      second: SingleType
+    ): SingleType {
+      if (first.expectedCppType != second.expectedCppType) {
+        throw IllegalArgumentException(
+          "Cannot merge types with different CppType: ${first.expectedCppType} and ${second.expectedCppType}"
+        )
+      }
+
+      val firstTypeParameters = first.parameterTypes
+      val secondTypeParameters = second.parameterTypes
+      if (firstTypeParameters == null || secondTypeParameters == null) {
+        return first
+      }
+
+      require(firstTypeParameters.size == secondTypeParameters.size) {
+        "Cannot merge types with different number of parameters: ${first.parameterTypes.size} and ${second.parameterTypes.size}"
+      }
+
+      val size = firstTypeParameters.size
+      val parameters = (0..<size).map { index ->
+        ExpectedType.merge(firstTypeParameters[index], secondTypeParameters[index])
+      }
+
+      return SingleType(
+        first.expectedCppType,
+        parameters.toTypedArray()
+      )
+    }
+  }
 }
 
 /**
@@ -63,11 +123,10 @@ class ExpectedType(
   @DoNotStrip
   fun getFirstType() = innerPossibleTypes.first()
 
-  operator fun plus(other: ExpectedType): ExpectedType {
-    return ExpectedType(
-      *this.innerPossibleTypes,
-      *other.innerPossibleTypes
-    )
+  override fun hashCode(): Int {
+    var result = innerCombinedTypes
+    result = 31 * result + innerPossibleTypes.contentHashCode()
+    return result
   }
 
   override operator fun equals(other: Any?): Boolean {
@@ -116,7 +175,8 @@ class ExpectedType(
     )
 
     fun fromKType(type: KType): ExpectedType {
-      val kClass = type.classifier as? KClass<*> ?: throw IllegalArgumentException("Cannot obtain KClass from '$type'")
+      val kClass = type.classifier as? KClass<*>
+        ?: throw IllegalArgumentException("Cannot obtain KClass from '$type'")
       when (kClass) {
         Int::class -> return ExpectedType(SingleType(CppType.INT))
         Long::class -> return ExpectedType(SingleType(CppType.LONG))
@@ -138,6 +198,20 @@ class ExpectedType(
         }
       }
       throw InvalidExpectedType(type)
+    }
+
+    fun merge(
+      vararg types: ExpectedType
+    ): ExpectedType {
+      val typesGroup = types
+        .flatMap { it.innerPossibleTypes.asIterable() }
+        .groupBy { it.expectedCppType }
+
+      val mergedTypes = typesGroup.map { (_, types) ->
+        types.reduce { a, b -> SingleType.merge(a, b) }
+      }
+
+      return ExpectedType(*mergedTypes.toTypedArray())
     }
   }
 }